#68 Refactor rename event_data lib to serialisation
