+++ nexus_file_reader/CMakeLists.txt
@@ -22,7 +22,7 @@ include_directories(
 add_library(nexusFileReader_lib ${SRC_FILES} ${INC_FILES})
 
 target_link_libraries(nexusFileReader_lib
-        eventdata_lib
+        serialisation_lib
         CONAN_PKG::hdf5
         CONAN_PKG::h5cpp
         CONAN_PKG::spdlog)
