+++ serialisation/include/DetectorSpectrumMapData.h
