+++ serialisation/src/DetectorSpectrumMapData.cpp
