+++ serialisation/src/SampleEnvironmentEventDouble.cpp
