+++ serialisation/test/DetectorSpectrumMapDataTest.cpp
